Drop support for ruby 1.8
Making this call, but @rubiii should feel free to disagree with me
if I'm in the wrong. ruby 1.8.7 has been EOL'd, and I think that I've
had a hard time keeping this up to date recently, but it's entirely
possible. This is the easiest thing that we can do that will keep
savon (and its associated projects (eg: gyoku)) green.