Allowing for order! to be a string key
When using savon to issue SOAP requests in a background process, the
request parameters (a hash) get saved by resque in redis. When pulled
out of redis, all keys of the hash are stringified as they are
marshalled as strings.
In order to avoid ordering issues when using resque or sidekick, gyoku
will accept both the symbol or the string version of the order! key and
respect that order.